% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mozzie}
\alias{mozzie}
\title{Weekly notified dengue cases in Sri Lanka}
\format{A data frame with 279 rows and 28 variables
\describe{
\item{ID}{row id}
\item{Year}{Year}
\item{Week}{Week number to locate the week within a year}
\item{Colombo}{Colombo district dengue cases}
\item{Gampaha}{Gampaha district dengue cases}
\item{Kalutara}{Kalutara district dengue cases}
\item{Kandy}{Kandy district dengue cases}
\item{Matale}{Matale district dengue cases}
\item{Nuwara Eliya}{Nuwara Eliya district dengue cases}
\item{Galle}{Galle district dengue cases}
\item{Hambantota}{Hambantota district dengue cases}
\item{Matara}{Matara district dengue cases}
\item{Jaffna}{Jaffna district dengue cases}
\item{Kilinochchi}{Kilinochchi district dengue cases}
\item{Mannar}{Mannar district dengue cases}
\item{Vavuniya}{Vavuniya district dengue cases}
\item{Mulative}{Mulative district dengue cases}
\item{Batticalo}{Batticalo district dengue cases}
\item{Ampara}{Ampara district dengue cases}
\item{Trincomalee}{Trincomalee district dengue cases}
\item{Kurunagala}{Kurunagala district dengue cases}
\item{Puttalam}{Puttalam district dengue cases}
\item{Anuradhapura}{Anuraghapura district dengue cases}
\item{Polonnaruwa}{Polonnaruwa district dengue cases}
\item{Badulla}{Badulla district dengue cases}
\item{Monaragala}{Monaragala district dengue cases}
\item{Ratnapura}{Ratnapura district dengue cases}
\item{Kegalle}{Kegalle district dengue cases}}}
\source{
weekly epidemiological reports published by the Epidemiological Unit, Ministry of Health, Sri Lanka
}
\usage{
mozzie
}
\description{
Weekly notified dengue cases in 25 districts of Sri Lanka from 2008/week-52 to 2014/week-21
}
\keyword{datasets}
